requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.11
python-dotenv==1.0.1
Werkzeug==3.1.3
zipp==3.21.0
